move all the individual query processors into the right folder
