+++ snuba/datasets/storages/errors_common.py
@@ -10,11 +10,6 @@ from snuba.clickhouse.columns import (
 from snuba.clickhouse.columns import SchemaModifiers as Modifiers
 from snuba.clickhouse.columns import String, UInt
 from snuba.datasets.storages.events_bool_contexts import EventsBooleanContextsProcessor
-from snuba.datasets.storages.processors.replaced_groups import (
-    PostReplacementConsistencyEnforcer,
-)
-from snuba.datasets.storages.type_condition_optimizer import TypeConditionOptimizer
-from snuba.datasets.storages.user_column_processor import UserColumnProcessor
 from snuba.query.conditions import ConditionFunctions, binary_condition
 from snuba.query.expressions import Column, Literal
 from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
@@ -26,8 +21,10 @@ from snuba.query.processors.empty_tag_condition_processor import (
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.prewhere import PrewhereProcessor
+from snuba.query.processors.replaced_groups import PostReplacementConsistencyEnforcer
 from snuba.query.processors.slice_of_map_optimizer import SliceOfMapOptimizer
 from snuba.query.processors.table_rate_limit import TableRateLimit
+from snuba.query.processors.type_condition_optimizer import TypeConditionOptimizer
 from snuba.query.processors.type_converters.hexint_column_processor import (
     HexIntColumnProcessor,
 )
@@ -40,6 +37,7 @@ from snuba.query.processors.type_converters.uuid_column_processor import (
 from snuba.query.processors.uniq_in_select_and_having import (
     UniqInSelectAndHavingProcessor,
 )
+from snuba.query.processors.user_column_processor import UserColumnProcessor
 from snuba.replacers.replacer_processor import ReplacerState
 from snuba.web.split import ColumnSplitQueryStrategy, TimeSplitQueryStrategy
