+++ snuba/datasets/storages/errors_v2.py
@@ -16,11 +16,6 @@ from snuba.datasets.storages.errors_common import (
     required_columns,
 )
 from snuba.datasets.storages.events_bool_contexts import EventsBooleanContextsProcessor
-from snuba.datasets.storages.processors.replaced_groups import (
-    PostReplacementConsistencyEnforcer,
-)
-from snuba.datasets.storages.type_condition_optimizer import TypeConditionOptimizer
-from snuba.datasets.storages.user_column_processor import UserColumnProcessor
 from snuba.datasets.table_storage import build_kafka_stream_loader_from_settings
 from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
     ArrayJoinKeyValueOptimizer,
@@ -32,9 +27,11 @@ from snuba.query.processors.empty_tag_condition_processor import (
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.prewhere import PrewhereProcessor
+from snuba.query.processors.replaced_groups import PostReplacementConsistencyEnforcer
 from snuba.query.processors.slice_of_map_optimizer import SliceOfMapOptimizer
 from snuba.query.processors.table_rate_limit import TableRateLimit
 from snuba.query.processors.tuple_unaliaser import TupleUnaliaser
+from snuba.query.processors.type_condition_optimizer import TypeConditionOptimizer
 from snuba.query.processors.type_converters.hexint_column_processor import (
     HexIntColumnProcessor,
 )
@@ -47,6 +44,7 @@ from snuba.query.processors.type_converters.uuid_column_processor import (
 from snuba.query.processors.uniq_in_select_and_having import (
     UniqInSelectAndHavingProcessor,
 )
+from snuba.query.processors.user_column_processor import UserColumnProcessor
 from snuba.replacers.replacer_processor import ReplacerState
 from snuba.subscriptions.utils import SchedulingWatermarkMode
 from snuba.utils.streams.topics import Topic
