+++ snuba/query/processors/type_condition_optimizer.py
