+++ tests/query/processors/test_type_condition_optimizer.py
@@ -1,7 +1,6 @@
 from snuba.clickhouse.columns import ColumnSet
 from snuba.clickhouse.formatter.expression import ClickhouseExpressionFormatter
 from snuba.clickhouse.query import Query
-from snuba.datasets.storages.type_condition_optimizer import TypeConditionOptimizer
 from snuba.query.conditions import (
     BooleanFunctions,
     ConditionFunctions,
@@ -9,6 +8,7 @@ from snuba.query.conditions import (
 )
 from snuba.query.data_source.simple import Table
 from snuba.query.expressions import Column, Literal
+from snuba.query.processors.type_condition_optimizer import TypeConditionOptimizer
 from snuba.query.query_settings import HTTPQuerySettings
